Making progress towards supporting signs.
Added signed flag to ast for reg declarations.
Added signed annotation to reg declaration when integers are lowered to regs in isolator.
Changed number flag from Format::UNSIGNED to Format::UNBASED for clarity.
Changed constructor in Number for case when Bits are used: no more size flag (it's encoded in the bits).
Logic for tracking sign changes in Bits.

Signed-off-by: eric schkufza <eric.schkufza@gmail.com>